+++ src/test/webapp/unit/checkBoxSpec.js
@@ -2,8 +2,7 @@ define(["unit/testUtils"], function(utils) {
 
     describe("checkbox", function() {
         it('should save the ui value into the model', function() {
-            // Note: Be sure to use ng:repeat, as this is the most problematic case!
-            var d = utils.compileInPage('<div ng:repeat="item in [1]">' +
+            var d = utils.compileInPage('<div>' +
                 '<input name="mysel" id="mysel" type="checkbox">{{mysel}}<label for="mysel" id="mylab">Entry</label>' +
                 '</div>');
             var page = d.page;
@@ -19,9 +18,8 @@ define(["unit/testUtils"], function(utils) {
         });
 
         it('should save the model value into the ui', function() {
-            // Note: Be sure to use ng:repeat, as this is the most problematic case!
             var d = utils.compileInPage(
-                '<div ng:repeat="item in [1]">' +
+                '<div>' +
                     '<input name="mysel" id="mysel" type="checkbox"><label for="mysel" id="mylab">Entry</label>' +
                     '</div>');
             var page = d.page;
@@ -39,9 +37,8 @@ define(["unit/testUtils"], function(utils) {
         });
 
         it('should use the disabled attribute', function() {
-            // Note: Be sure to use ng:repeat, as this is the most problematic case!
             var d = utils.compileInPage(
-                '<div ng:repeat="item in [1]">' +
+                '<div>' +
                     '<input name="mysel" id="mysel" type="checkbox" value="false" ng:bind-attr="{disabled: \'{{disabled}}\'}"><label for="mysel" id="mylab">Entry</label>' +
                     '</div>');
             var page = d.page;
@@ -54,19 +51,6 @@ define(["unit/testUtils"], function(utils) {
             scope.$set('disabled', true);
             scope.$eval();
             expect(parentDiv.hasClass('ui-disabled')).toBeTruthy();
-        });
-
-        it('should enhance checkboxes when ng:repeat is used (not only during page compile)', function() {
-            var d = utils.compileInPage(
-                '<div><div ng:repeat="item in items" id="mydiv">' +
-                    '<input name="mysel" id="mysel{{$index}}" type="checkbox" value="false"><label for="mysel{{$index}}" id="mylab{{$index}}">Entry</label>' +
-                    '</div></div>');
-            var page = d.page;
-            var parentDiv = d.element;
-            expect(parentDiv.find('.ui-checkbox').length).toBe(0);
-            page.scope().$set('items', [1,2]);
-            page.scope().$eval();
-            expect(parentDiv.find('.ui-checkbox').length).toBe(2);
         });
     });