+++ src/test/webapp/unit/radioSpec.js
@@ -3,8 +3,7 @@ define(["unit/testUtils"], function(utils) {
     describe("radio", function() {
 
         it('should save the ui value into the model', function() {
-            // Note: Be sure to use ng:repeat, as this is the most problematic case!
-            var d = utils.compileInPage('<div ng:repeat="item in [1]">' +
+            var d = utils.compileInPage('<div>' +
                 '<input name="mysel" id="mysel" type="radio" value="v1"><label for="mysel" id="mylab">Entry</label>' +
                 '</div>');
             var input = d.element.find("input");
@@ -18,8 +17,7 @@ define(["unit/testUtils"], function(utils) {
 
         });
         it('should save the model value into the ui', function() {
-            // Note: Be sure to use ng:repeat, as this is the most problematic case!
-            var d = utils.compileInPage('<div ng:repeat="item in [1]">' +
+            var d = utils.compileInPage('<div>' +
                 '<input name="mysel" id="mysel" type="radio" value="v1"><label for="mysel" id="mylab">Entry</label>' +
                 '</div>');
             var input = d.element.find("input");
@@ -36,9 +34,8 @@ define(["unit/testUtils"], function(utils) {
         });
 
         it('should use the disabled attribute', function() {
-            // Note: Be sure to use ng:repeat, as this is the most problematic case!
             var d = utils.compileInPage(
-                '<div ng:repeat="item in [1]">' +
+                '<div>' +
                     '<input name="mysel" id="mysel" type="radio" value="v1" ng:bind-attr="{disabled: \'{{disabled}}\'}"><label for="mysel" id="mylab">Entry</label>' +
                     '</div>');
             var input = d.element.find("input");