+++ src/test/webapp/unit/testUtils.js
@@ -23,7 +23,7 @@ define(["jquery", "angular"], function($, angular) {
 
     function compileInPage(html) {
         var elements = test$("<div>"+html+"</div>").children().addClass("result", "true");
-        var page = test$('<div id="start" data-role="page"><div data-role="content"></div></div>');
+        var page = test$('<div id="start" data-role="page" data-url="start"><div data-role="content"></div></div>');
         test$("body").append(page);
         page.find(":jqmData(role='content')").append(elements);
         page.page();