+++ src/test/webapp/unit/textInputSpec.js
@@ -3,8 +3,7 @@ define(["unit/testUtils"], function(utils) {
     describe("textInput", function() {
 
         it('should save the ui value into the model', function() {
-            // Note: Be sure to use ng:repeat, as this is the most problematic case!
-            var d = utils.compileInPage('<input ng:repeat="item in [1]" name="mysel" id="mysel" type="text">');
+            var d = utils.compileInPage('<input name="mysel" type="text">');
             var input = d.element;
             var scope = input.scope();
             expect(scope.$get('mysel')).toBeFalsy();
@@ -14,8 +13,7 @@ define(["unit/testUtils"], function(utils) {
         });
 
         it('should save the model value into the ui', function() {
-            // Note: Be sure to use ng:repeat, as this is the most problematic case!
-            var d = utils.compileInPage('<input ng:repeat="item in [1]" name="mysel" id="mysel" type="text">');
+            var d = utils.compileInPage('<input name="mysel" type="text">');
             var input = d.element;
             var scope = input.scope();
             expect(input[0].value).toEqual('');
@@ -25,8 +23,7 @@ define(["unit/testUtils"], function(utils) {
         });
 
         it('should use the disabled attribute', function() {
-            // Note: Be sure to use ng:repeat, as this is the most problematic case!
-            var d = utils.compileInPage('<input ng:repeat="item in [1]" name="mysel" id="mysel" type="text" ng:bind-attr="{disabled: \'{{disabled}}\'}">');
+            var d = utils.compileInPage('<input name="mysel" type="text" ng:bind-attr="{disabled: \'{{disabled}}\'}">');
             var input = d.element;
             var scope = input.scope();
             scope.$set('disabled', false);
@@ -37,20 +34,8 @@ define(["unit/testUtils"], function(utils) {
             expect(input.hasClass('ui-disabled')).toBeTruthy();
         });
 
-        it('should be removable when ng:repeat shrinks', function() {
-            var d = utils.compileInPage('<div ng:init="mylist = [1,2]">' +
-                '<input ng:repeat="item in mylist" name="mysel" id="mysel" type="text">' +
-                '</div>');
-            var container = d.element;
-            var scope = container.scope();
-            expect(container.children('input').length).toEqual(2);
-            scope.mylist = [1];
-            scope.$eval();
-            expect(container.children('input').length).toEqual(1);
-        });
-
         it('should work with type="tel"', function() {
-            var d = utils.compileInPage('<input ng:repeat="item in [1]" name="mysel" id="mysel" type="tel">');
+            var d = utils.compileInPage('<input name="mysel" type="tel">');
             var input = d.element;
             expect(input.prop('type')).toEqual('tel');
             var scope = input.scope();
@@ -61,7 +46,7 @@ define(["unit/testUtils"], function(utils) {
         });
 
         it('should work with type="number"', function() {
-            var d = utils.compileInPage('<input ng:repeat="item in [1]" name="mysel" id="mysel" type="number">');
+            var d = utils.compileInPage('<input name="mysel" type="number">');
             var input = d.element;
             expect(input.prop('type')).toEqual('number');
             var scope = input.scope();